Edited popup messages for poslih language only Fixed bug with clearing list of users flats Fixed calculate trend bug with values uder -10%
Added authentication of user when deleting transaction from DB

Added edit option of existing flat with user authentication

+++ web/src/main/java/com/infoshareacademy/jbusters/dao/NewTransactionDao.java
@@ -33,7 +33,7 @@ public class NewTransactionDao {
         }
     }
 
-    public NewTransaction findById(Long id) {
+    public NewTransaction findById(int id) {
         return entityManager.find(NewTransaction.class, id);
     }
